% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{getSquareOf}
\alias{getSquareOf}
\title{Calculate the Square of a Variable}
\usage{
getSquareOf(x)
}
\arguments{
\item{x}{a numeric}
}
\value{
a numeric
}
\description{
Blacalgkjigjisghldrijgldrij
}
\author{
M. Fortin
}
