% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{getSquareRootOf}
\alias{getSquareRootOf}
\title{Calculate the Square Root of a Variable}
\usage{
getSquareRootOf(x)
}
\arguments{
\item{x}{a numeric}
}
\value{
a numeric
}
\description{
Blacalgkjigjisghldrijgldrij
}
\author{
M. Fortin
}
